🗃 All in main.rs for later restructuring
