--- build.rs
@@ -1,4 +1,0 @@
-fn main() {
-    #[cfg(all(windows, target_env = "msvc"))]
-    std::env::set_var("FFMPEG_DIR", "./ffmpeg");
-}
