--- src/actions/mod.rs
@@ -1,5 +1,0 @@
-pub mod capture;
-pub mod play;
-
-pub use capture::*;
-pub use play::*;
