--- src/opengl/mod.rs
@@ -1,8 +1,0 @@
-pub mod context;
-pub mod drawing;
-pub mod layout;
-pub mod notes;
-pub mod objects;
-pub mod particles;
-pub mod shaders;
-pub mod uniforms;
