--- src/ui/mod.rs
@@ -1,3 +1,0 @@
-pub mod cli;
-pub mod gui;
-pub mod theme;
